Add like counts and unit test
