+++ README.md
@@ -1,3 +1,63 @@
-# Pokemon
-<!-- https://css-tricks.com/how-to-create-neon-text-with-css/ -->
-<!-- https://epiktheme.com/demos/html/Gamix-Preview/demos/index3.html -->
+![](https://img.shields.io/badge/Microverse-blueviolet)
+
+# Pokilo
+Pokilo displays information about the Pokemons for which user can post likes and comments on each pokemon.
+
+# Live Demo
+[Live Demo](https://riyabulia12.github.io/Pokilo/dist/)
+
+## Built With
+- HTML5, CSS3, ES6, Webpack, HtmlWebpackPlugin, Jest, REST API
+
+## Project Setup
+To get a local copy up and running follow these simple example steps.
+
+### Prerequisites
+
+- You need to have Code Editor Installed
+- Make sure Nodejs and NPM is installed in your system for linters and libraries
+
+### Setup
+1. Clone the repo in your local machine:
+```bash
+git clone https://github.com/RiyaBulia12/Pokilo.git
+```
+2. Open terminal and cd into project and install the dependencies
+```bash
+cd Pokilo
+npm install
+```
+
+3. run the command to open your index.html automatically in your brower:
+```bash
+npm run build
+npm start
+```
+
+## Authors
+
+👤 **Riya Bulia**
+
+- GitHub: [@RiyaBulia12](https://github.com/RiyaBulia12)
+- Twitter: [@Rbulia_original](https://twitter.com/@rbulia_original)
+- LinkedIn: [Riya-Bulia](https://linkedin.com/in/riya-bulia)
+
+👤 **Recilla Khamala**
+
+- GitHub: [@Rn486](https://github.com/Rn486)
+- Twitter: [@recillahk](https://twitter.com/recillahk)
+- LinkedIn: [LinkedIn](https://www.linkedin.com/in/recillah-khamala-071151b7/)
+
+
+## 🤝 Contributing
+
+Contributions, issues, and feature requests are welcome!
+
+## Show your support
+
+Give a ⭐️ if you like this project!
+
+## Acknowledgments
+- [PokeAPI](https://pokeapi.co/)
+- [Create neon text using CSS](https://css-tricks.com/how-to-create-neon-text-with-css/)
+- [UI Inspired from Gamix template](https://epiktheme.com/demos/html/Gamix-Preview/demos/index3.html)
