+++ mocks/fileMock.js
@@ -1,0 +1,2 @@
+// Mocking import image file
+export default '';
