--- src/assets/js/functions.js
@@ -1,9 +1,0 @@
-export function importImages(r) {
-  const images = {};
-  r.keys().forEach((item) => {
-    images[item.replace('./', '')] = r(item);
-  });
-  return images;
-}
-
-export default { importImages };
