+++ src/index.html
@@ -15,18 +15,18 @@
    <div class="container-fluid">
       <header class="header">
          <nav class="menu-container navbar fixed-top navbar-light p-0">
-            <div class="logo">
-               <a class="logo-img" href="#">
-                  <img src="../src/assets/img/logo.png" alt="Logo Icon"> </a>
+            <div class="logo d-flex ms-sm-5 ">
+               <a class="logo-text" href="#"> POKÍLÖ <span class="logo-dot">.</span>
+               </a>
             </div>
-            <div class="container " id="menu-list">
-               <ul class="nav my-4 fs-6 text-uppercase" id="nav">
-                  <li class="nav-item px-5"><a href="#pokemon" class="nav-link" aria-label="Pokemon">Pokemon <span
+            <div class="p-0" id="menu-list">
+               <ul class="nav my-md-4 fs-6 text-uppercase" id="nav">
+                  <li class="nav-item px-md-5"><a href="#pokemon" class="nav-link" aria-label="Pokemon">Pokemon <span
                            class="poke-counts" id="pokeCounts"></span>
                      </a></li>
-                  <li class="nav-item px-5"><a href="#pokemon" class="nav-link" aria-label="Pokemon">Video Games</a>
+                  <li class="nav-item px-md-5 "><a href="#pokemon" class="nav-link" aria-label="Pokemon">Video Games</a>
                   </li>
-                  <li class="nav-item px-5"><a href="#pokemon" class="nav-link" aria-label="Pokemon">Events</a></li>
+                  <li class="nav-item px-md-5"><a href="#pokemon" class="nav-link" aria-label="Pokemon">Events</a></li>
                </ul>
             </div>
          </nav>
@@ -55,7 +55,7 @@
          </div>
       </div>
       <footer class="footer">
-         <p>Created by Recillah and Riya under the Microverse guidance</p>
+         <p class="m-2">Created by Recillah and Riya under the Microverse guidance</p>
       </footer>
    </div>
    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
